+++ NEWS.md
@@ -2,6 +2,8 @@
 * Default rats method improved for multidimensional datasets.
 * Added norm_mrs function to normalise the intenisty of spectral data.
 * Added bc_constant function to correct spectral baselines by a constant offset.
+* Added re_weighting function to apply a resolution enhancement weighting to the
+FID.
 
 # spant 0.15.0
 * Bug fix for using auto_phase function with a single spectrum.